service/requirements.txt
fastapi
python-telegram-bot
python-multipart
uvicorn